+++ src/components/SidePanel.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import LoginButton from "../atoms/LoginButton";
-import SetEvent from "../atoms/SetEvent";
+import SetEvent from "./SetEvent";
 
 const SidePanel = () => {
     const userProfile = useSelector(state => state.userProfile);
